Separate PriorityNode from PriorityQueue
